Drop NAE name space
